+++ src/device.h
@@ -23,7 +23,7 @@
   #include "port.h"
   #include "mem.h"
   #include "queue.h"
-  #include "stream_helios.h"
+  #include "streams.h"
   #include "sys.h"
   #include "task.h"
   #include "timer.h"
