+++ src/timer.h
@@ -24,7 +24,7 @@
   #include "device.h"
   #include "mem.h"
   #include "queue.h"
-  #include "stream_helios.h"
+  #include "streams.h"
   #include "sys.h"
   #include "task.h"
